Optimize solver hot path round 2: cached departureTime, zero-alloc ArbZG, two-level routing cache
1. Cache departureTime as derived field in Event
   - Computed in setArrivalTime(), returned directly from getDepartureTime()
   - Eliminates ~20K+ Instant allocations/sec in solver loop
   - arrivalTime is only mutated by VariableListener, so setter caching is complete

2. Array sort replaces stream+sort+collect in consecutive driving constraint
   - calculateConsecutiveDrivingPenalty uses Event[] + Arrays.sort
   - Eliminates stream pipeline + ArrayList allocation per evaluation

3. For-loop with epoch-second arithmetic replaces streams in ArbZG constraints
   - maxDailyWorkingHours and maxWeeklyWorkingHours use direct
     (departure.epochSecond - arrival.epochSecond) / 60
   - Eliminates Duration.between() object allocation per event per evaluation

4. Two-level ConcurrentHashMap replaces LocationPair record cache
   - ConcurrentHashMap<Location, ConcurrentHashMap<Location, TravelData>>
   - Eliminates LocationPair allocation on every cache lookup (~10-20K/sec)
   - LocationPair record removed entirely

5. Minor hot-path fixes
   - vehicleCapacityConstraint: local variable for cumulativePassengerCount (3 calls → 1)
   - timeWindowConstraint: plusSeconds() replaces Duration.ofMinutes()
   - Driver.maxCapacity: long → int (avoids widening conversions)
   - getEventDate(): removed dead fallback branches (shiftDate always set)
   - WebResource.shiftsPage(): groupingBy replaces 7 stream filters

All changes provably isomorphic: same outputs given same inputs.

+++ src/main/java/com/vrp/domain/Driver.java
@@ -9,7 +9,7 @@ public class Driver implements Standstill {
     
     private String id;
     private Location homeLocation;
-    private long maxCapacity = 6;
+    private int maxCapacity = 6;
     private Duration maxConsecutiveHours = Duration.ofHours(4);
     private Duration minBreak = Duration.ofMinutes(30);
     private Duration maxBreak = Duration.ofHours(4);
@@ -57,11 +57,11 @@ public class Driver implements Standstill {
         this.homeLocation = homeLocation;
     }
     
-    public long getMaxCapacity() {
+    public int getMaxCapacity() {
         return maxCapacity;
     }
-    
-    public void setMaxCapacity(long maxCapacity) {
+
+    public void setMaxCapacity(int maxCapacity) {
         this.maxCapacity = maxCapacity;
     }
